eco-task/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.0
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.9
django-bulk-update==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
dnspython==2.1.0
eventlet==0.33.0
greenlet==1.1.2
idna==3.3
importlib-metadata==4.8.2
kombu==5.2.2
Markdown==3.3.6
prompt-toolkit==3.0.23
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.0.2
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.6.0
